feat: check user is a team admin before deleting
[Ticket: FYP-38]

+++ src/routes/project/project.js
@@ -73,20 +73,27 @@ module.exports = function (app, db, admin) {
       if (!doc.exists) {
         return res.status(404).json({ error: "No such document" });
       }
-      await docRef.delete().then((response) => {
-        // Delete project ID from team
-        const teamDocRef = db.collection("teams").doc(doc.data().teamId);
-        teamDocRef
-          .update({
-            teamProjects: admin.firestore.FieldValue.arrayRemove(docRef.id),
-          })
-          .then(() => {
-            return res.json(response);
-          })
-          .catch((error) => {
-            return res.status(500).json(error);
-          });
-      });
+      const teamDocRef = db.collection("teams").doc(doc.data().teamId);
+      const teamDoc = await teamDocRef.get();
+
+      // Check is user is admin of project's team first
+      if(teamDoc.data().teamAdmins.includes(req.uid)){
+        await docRef.delete().then((response) => {
+          // Delete project ID from team
+          const teamDocRef = db.collection("teams").doc(doc.data().teamId);
+          teamDocRef
+            .update({
+              teamProjects: admin.firestore.FieldValue.arrayRemove(docRef.id),
+            })
+            .then(() => {
+              return res.json(response);
+            })
+            .catch((error) => {
+              return res.status(500).json(error);
+            });
+        });
+      }
+      return res.status(403).json({ error: "Only team admins can delete projects" })
     });
 
   app.route("/api/getAllProjectsForUser").get(async function (req, res) {
